Fixinf

+++ Ponyliga/Ponyliga/Views/Admin/TeamDeletionPage.xaml.cs
@@ -59,19 +59,23 @@ namespace Ponyliga.Views.Admin
 
         public void btn_deleteTeam_Clicked(object sender, EventArgs e)
         {
-            string name = TeamPicker.Items[TeamPicker.SelectedIndex];
+            if(TeamPicker.SelectedIndex != -1)
+            {
+                string name = TeamPicker.Items[TeamPicker.SelectedIndex];
 
-            Team team = new Team();
+                Team team = new Team();
 
-            int teamId = taskTeam.Find(t => t.name == name).id;
-            team.id = teamId;
+                int teamId = taskTeam.Find(t => t.name == name).id;
+                team.id = teamId;
 
-            ApiService apiService = new ApiService();
-            apiService.DeleteTeam(team.id.ToString());
+                ApiService apiService = new ApiService();
+                apiService.DeleteTeam(team.id.ToString());
 
-            DisplayAlert("Warnung", "Team wird gelöscht!", "OK");
-            FillTeamList();
-            TeamPicker.SelectedIndex = -1;
+                DisplayAlert("Warnung", "Team wird gelöscht!", "OK");
+                FillTeamList();
+                TeamPicker.SelectedIndex = -1;
+            }
+            
         }
 
     }
